test(testing): ArticleScreen tests вынесены в отдельные suite files
state, content rendering, controller orchestration и navigation tests для ArticleScreen выделены из RSSReaderTests.swift в самостоятельные feature-oriented test suites
