+++ README.md
@@ -328,6 +328,7 @@
 - [x] вынести общий test-support и sync/cross-device tests из `RSSReaderTests.swift` в отдельные файлы, чтобы дальнейшая декомпозиция test target шла по feature-oriented структуре;
 - [x] перейти от `extension RSSReaderTests` к standalone suite types и начать разнос service-level тестов по feature-oriented файлам (`Sync`, `Feeds`, `Articles`), чтобы navigator и test target отражали логические границы ответственности;
 - [x] вынести `Sidebar` query/presentation/state tests из `RSSReaderTests.swift` в отдельные suite files, чтобы feature-oriented структура test target отражала отдельные read-model, presentation и screen-state обязанности;
+- [x] вынести `ArticleScreen` state/rendering/controller/navigation tests из `RSSReaderTests.swift` в отдельные suite files, чтобы тесты reader detail flow были разложены по screen state, content rendering, controller orchestration и navigation policy;
 - [ ] зафиксировать целевой sync scope: в CloudKit должны синхронизироваться `Feed`, `Folder`, `ArticleState` и `AppSettings`, а `Article` и `FeedFetchLog` должны оставаться только локальным storage слоем;
 - [ ] провести audit CloudKit compatibility для `AppSettings`, `Feed` и `Folder`: проверить `@Attribute(.unique)`, relationship semantics и текущие invariants, которые могут блокировать SwiftData sync;
 - [ ] провести audit CloudKit compatibility для `ArticleState`, `Article` и `FeedFetchLog`: проверить `#Unique`, nonoptional relationships, delete rules и закрепить границу между sync-backed сущностями и локальным article cache;
