+++ RSSReaderTests/ArticleScreen/ArticleScreenContentRendererTests.swift
@@ -1,0 +1,216 @@
+import Foundation
+import Testing
+@testable import RSSReader
+
+@Suite("Article Screen / Content Rendering")
+@MainActor
+struct ArticleScreenContentRendererTests {
+    @Test
+    func articleScreenContentHeaderFormatsFieldsInPublishedTitleAuthorFeedOrder() {
+        let publishedAt = Date(timeIntervalSince1970: 1_700_000_000)
+        let content = ArticleScreenContentState(
+            article: makeReaderArticleDTO(
+                feedTitle: "THECODE.MEDIA",
+                author: "Юлия Зубарева",
+                publishedAt: publishedAt
+            )
+        )
+
+        #expect(content.header.publishedAtText == ArticleScreenDateFormatter.string(from: publishedAt))
+        #expect(content.header.title == "Article")
+        #expect(content.header.author == "Юлия Зубарева")
+        #expect(content.header.feedTitle == "THECODE.MEDIA")
+    }
+
+    @Test
+    func articleScreenContentHeaderHidesBlankMetadataAndFallsBackForBlankTitle() {
+        let content = ArticleScreenContentState(
+            article: makeReaderArticleDTO(
+                feedTitle: "   ",
+                title: "   ",
+                author: " \n ",
+                publishedAt: nil
+            )
+        )
+
+        #expect(content.header.publishedAtText == nil)
+        #expect(content.header.title == "Untitled Article")
+        #expect(content.header.author == nil)
+        #expect(content.header.feedTitle == nil)
+    }
+
+    @Test
+    func articleScreenStateUsesExistingRenderingPriorityForBodyContent() {
+        var state = ArticleScreenState()
+        let article = makeReaderArticleDTO(
+            summary: "Summary copy",
+            contentHTML: "<p>HTML body</p>",
+            contentText: "Longer content text"
+        )
+
+        state.applyLoadedArticle(article)
+
+        #expect(state.derivedViewState().content?.body.blocks == [.paragraph(.plainText("HTML body"))])
+        #expect(state.derivedViewState().content?.body.source == .contentHTML)
+    }
+
+    @Test
+    func articleScreenContentRendererParsesHTMLParagraphsAndInlineImagesInOrder() {
+        let content = ArticleScreenContentState(
+            article: makeReaderArticleDTO(
+                summary: "Short summary",
+                contentHTML: """
+                <p>First paragraph.</p>
+                <img src="https://example.com/images/inline.png">
+                <p>Second <strong>paragraph</strong>.</p>
+                """,
+                contentText: "Plain text fallback"
+            )
+        )
+
+        #expect(
+            content.body.blocks == [
+                .paragraph(.plainText("First paragraph.")),
+                .image(URL(string: "https://example.com/images/inline.png")!),
+                .paragraph(.plainText("Second paragraph."))
+            ]
+        )
+        #expect(content.body.source == .contentHTML)
+    }
+
+    @Test
+    func articleScreenContentRendererPreservesAnchorMetadataInsideHTMLParagraphs() {
+        let content = ArticleScreenContentState(
+            article: makeReaderArticleDTO(
+                contentHTML: """
+                <p>Read <a href="/guides/swift">Swift Guide</a> today.</p>
+                """,
+                canonicalURL: "https://example.com/articles/body"
+            )
+        )
+
+        #expect(
+            content.body.blocks == [
+                .paragraph(
+                    ArticleScreenTextBlock(
+                        spans: [
+                            ArticleScreenTextSpan(text: "Read "),
+                            ArticleScreenTextSpan(
+                                text: "Swift Guide",
+                                linkURL: URL(string: "https://example.com/guides/swift")!
+                            ),
+                            ArticleScreenTextSpan(text: " today.")
+                        ]
+                    )
+                )
+            ]
+        )
+        #expect(content.body.source == .contentHTML)
+    }
+
+    @Test
+    func articleScreenContentRendererDetectsLinksInsidePlainTextBody() {
+        let content = ArticleScreenContentState(
+            article: makeReaderArticleDTO(
+                contentText: "Read more at https://example.com/guides/swift today."
+            )
+        )
+
+        #expect(
+            content.body.blocks == [
+                .paragraph(
+                    ArticleScreenTextBlock(
+                        spans: [
+                            ArticleScreenTextSpan(text: "Read more at "),
+                            ArticleScreenTextSpan(
+                                text: "https://example.com/guides/swift",
+                                linkURL: URL(string: "https://example.com/guides/swift")!
+                            ),
+                            ArticleScreenTextSpan(text: " today.")
+                        ]
+                    )
+                )
+            ]
+        )
+        #expect(content.body.source == .contentText)
+    }
+
+    @Test
+    func articleScreenTextBlockBuildsAttributedStringWithLinkAttributes() {
+        let textBlock = ArticleScreenTextBlock(
+            spans: [
+                ArticleScreenTextSpan(text: "Read "),
+                ArticleScreenTextSpan(
+                    text: "Swift Guide",
+                    linkURL: URL(string: "https://example.com/guides/swift")!
+                ),
+                ArticleScreenTextSpan(text: " today.")
+            ]
+        )
+
+        let attributedString = textBlock.attributedString
+
+        #expect(String(attributedString.characters) == "Read Swift Guide today.")
+
+        let linkRuns = attributedString.runs.filter { $0.link != nil }
+        #expect(linkRuns.count == 1)
+        #expect(linkRuns.first?.link == URL(string: "https://example.com/guides/swift")!)
+        #expect(String(attributedString[linkRuns[0].range].characters) == "Swift Guide")
+    }
+
+    @Test
+    func articleScreenContentRendererUsesSummaryWithFallbackNoticeWhenFullBodyIsUnavailable() {
+        let content = ArticleScreenContentState(
+            article: makeReaderArticleDTO(
+                summary: """
+                Short summary paragraph.
+
+                Another summary paragraph.
+                """,
+                contentHTML: nil,
+                contentText: nil
+            )
+        )
+
+        #expect(
+            content.body.blocks == [
+                .paragraph(.plainText("Short summary paragraph.")),
+                .paragraph(.plainText("Another summary paragraph.")),
+                .fallbackNotice("This source only provides a summary, not the full article body.")
+            ]
+        )
+        #expect(content.body.source == .summary)
+    }
+
+    @Test
+    func articleScreenContentRendererBuildsGracefulFallbackWhenFeedHasNoBodyContent() {
+        let content = ArticleScreenContentState(
+            article: makeReaderArticleDTO(
+                summary: nil,
+                contentHTML: nil,
+                contentText: nil
+            )
+        )
+
+        #expect(
+            content.body.blocks == [
+                .fallbackNotice("Full article content is unavailable in this feed.")
+            ]
+        )
+        #expect(content.body.source == .empty)
+        #expect(content.body.readerMode == .embedded)
+    }
+
+    @Test
+    func articleScreenBodyContentStateDefinesFutureFullTextExtensionPoint() {
+        let extractedContent = ArticleScreenBodyContentState.extractedFullText(
+            blocks: [
+                .paragraph(.plainText("Extracted full text paragraph."))
+            ]
+        )
+
+        #expect(extractedContent.source == .fullTextExtracted)
+        #expect(extractedContent.readerMode == .fullText)
+        #expect(extractedContent.blocks == [.paragraph(.plainText("Extracted full text paragraph."))])
+    }
+}
